+++ zorka-core/src/main/java/com/jitlogic/zorka/core/perfmon/ThreadMonitorItem.java
@@ -101,4 +101,9 @@ public class ThreadMonitorItem {
     public String getName() {
         return name;
     }
+
+    @Override
+    public String toString() {
+        return "ITM(" + id + ", '" + name + "')";
+    }
 }